+++ examples/chat/manage.py
@@ -21,63 +21,24 @@ try:
 except ImportError: #pragma nocover
     sys.path.append('../../')
     import pulsar
-from pulsar.apps import ws, wsgi, rpc
+from pulsar.apps import ws, wsgi, rpc, pubsub
 
 CHAT_DIR = os.path.dirname(__file__)
 
-################################################################################
-##    Publish and clients methods
-def publish(message):
-    message = {'time': time.time(), 'message': message}
-    pulsar.send('webchat', 'publish_message', json.dumps(message))
-
-def broadcast(message):
-    remove = set()
-    clients = get_clients()
-    for client in clients:
-        try:
-            client.current_consumer.write(message)
-        except Exception:
-            remove.add(client)
-    clients.difference_update(remove)
-    
-def get_clients():
-    actor = pulsar.get_actor()
-    clients = actor.params.clients
-    if clients is None:
-        clients = set()
-        actor.params.clients = clients
-    return clients
-
-################################################################################
-##    Internal message passing
-@pulsar.command(ack=False)
-def publish_message(request, message):
-    monitor = request.actor
-    if monitor.managed_actors:
-        for worker in monitor.managed_actors.values():
-            monitor.send(worker, 'broadcast_message', message)
-    else:
-        broadcast(message)
-        
-@pulsar.command(ack=False)
-def broadcast_message(request, message):
-    broadcast(message)
-
-################################################################################
-##    MIDDLEWARE
+##    RPC MIDDLEWARE To publish messages
 class Rpc(rpc.PulsarServerCommands):
     
     def rpc_message(self, request, message):
-        publish(message)
+        pubsub.publish(message)
         return 'OK'
     
-    
+
+##    Web Socket Chat handler
 class Chat(ws.WS):
-        
+    
     def on_open(self, request):
         # Add pulsar.connection environ extension to the set of active clients
-        get_clients().add(request.cache['websocket'])
+        pubsub.subscribe(request.cache['websocket'])
         
     def on_message(self, request, msg):
         if msg:
@@ -88,24 +49,32 @@ class Chat(ws.WS):
                     lines.append(l)
             msg = ' '.join(lines)
             if msg:
-                publish(msg)
+                pubsub.publish(msg)
 
 
-def page(environ, start_response):
-    """ This resolves to the web page or the websocket depending on the path."""
-    path = environ.get('PATH_INFO')
-    if not path or path == '/':
+class WebChat(wsgi.LazyWsgi):
+    
+    def __init__(self, name):
+        self.name = name
+        
+    def setup(self):
+        # Register a pubsub handler
+        pubsub.register_handler(pubsub.PulsarPubSub(self.name))
+        return wsgi.WsgiHandler([wsgi.Router('/', get=self.home_page),
+                                 ws.WebSocket('/message', Chat()),
+                                 wsgi.Router('/rpc', post=Rpc())])
+        
+    def home_page(self, request):
         data = open(os.path.join(CHAT_DIR, 'chat.html')).read()
-        data = data % environ
-        start_response('200 OK', [('Content-Type', 'text/html'),
-                                  ('Content-Length', str(len(data)))])
-        return [pulsar.to_bytes(data)]
+        request.response.content_type = 'text/html'
+        request.response.content = data % request.environ
+        return request.response.start()
 
-def server(**kwargs):
-    chat = ws.WebSocket('/message', Chat())
-    api = rpc.RpcMiddleware(Rpc(), path='/rpc')
-    middleware = wsgi.WsgiHandler(middleware=(chat, api, page))
-    return wsgi.WSGIServer(name='webchat', callable=middleware, **kwargs)
+
+def server(callable=None, name=None, **kwargs):
+    name = name or 'webchat'
+    chat = WebChat(name)
+    return wsgi.WSGIServer(name=name, callable=chat, **kwargs)
 
 
 if __name__ == '__main__':  #pragma nocover